+++ app/controllers/public/pages_controller.rb
@@ -3,9 +3,12 @@
 module Public
   class PagesController < BaseController
     def home
-      return if current_organisation && current_user&.organisation != current_organisation
+      return if current_organisation.present?
+      return redirect_to new_user_session_path if current_user.blank?
+      return redirect_to manage_root_path if current_user.role_manager?
+      return redirect_to admin_root_path if current_user.role_admin?
 
-      redirect_to redirection_target
+      raise ActionController::RoutingError
     end
 
     def ext
@@ -16,12 +19,8 @@ module Public
 
     protected
 
-    def redirection_target
-      return new_user_session_path if current_organisation.blank? && current_user.blank?
-      return admin_root_path if current_user.role_admin?
-      return manage_root_path if current_user.role_manager?
-
-      '/404'
+    def current_organisation
+      @current_organisation ||= Organisation.find_by(slug: params[:org].presence)
     end
   end
 end