+++ app/params/manage/user_params.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Manage
+  class UserParams < ApplicationParams
+    def self.permitted_keys
+      %i[email role]
+    end
+  end
+end