+++ db/migrate/20200925123740_move_all_notification_markdown_templates_into_namespace.rb
@@ -1,0 +1,7 @@
+class MoveAllNotificationMarkdownTemplatesIntoNamespace < ActiveRecord::Migration[6.0]
+  def up
+    MarkdownTemplate.where(MarkdownTemplate.arel_table[:key].matches('%_message')).find_each do |template|
+      template.update(key: template.key.chomp('_message'), namespace: :notification)
+    end
+  end
+end