+++ CHANGELOG.md
@@ -1,21 +1,30 @@
 # Changelog
 
-## 20.9.2
+## [Unreleased]
+
+## [20.10.1] - 2020-10-01
+
+- Improve user management
+- Improve development setup
+
+## [20.9.2] - 2020-09-24
 
 - Update dependencies
 - Improve multi organisation support
+- Add namespace support for templates
+- Release code as open-source
 
-## 20.9.1
+## [20.9.1] 2020-09-04
 
 - Improve multi organisation support
 - Improve locale support
 
-## 20.8.5
+## [20.8.5] 2020-08-30
 
 - Improved error pages
 - Changed behaviour after login
 - Bugfixes
 
-## 20.8.4
+## [20.8.4] - 2020-08-25
 
 - Bugfixes